requirements.txt
numpy
pandas
scikit-learn
xgboost
matplotlib
seaborn
streamlit
pyyaml
